+++ src/bin/lsystem/describe.rs
@@ -1,5 +1,5 @@
 use std::process::ExitCode;
-use rusty_systems::system::family::abop;
+use rusty_systems::interpretation::abop;
 use rusty_systems::system::family::TokenDescription;
 use crate::heading_style;